+++ src/config.py
@@ -1,0 +1,32 @@
+"""Configuration and shared state."""
+
+__version__ = "0.0.3"
+
+import asyncio
+import os
+import uuid
+
+from dotenv import load_dotenv
+
+load_dotenv()
+
+CONFIG = {
+    "port": 3000,
+    "session_id": str(uuid.uuid4()),
+    "check_interval": 30 * 60,  # 30 minutes in seconds
+    "autonomous_mode": True,
+    "discord_webhook_url": os.getenv("DISCORD_WEBHOOK_URL", ""),  # Set via environment variable
+    "github_repo": os.getenv("GITHUB_REPO", ""),  # e.g. "suojae/smol-claw"
+    "github_webhook_secret": os.getenv("GITHUB_WEBHOOK_SECRET", ""),
+    "usage_limits": {
+        "max_calls_per_minute": 5,
+        "max_calls_per_hour": 20,
+        "max_calls_per_day": 500,
+        "min_call_interval_seconds": 5,
+        "warning_threshold_pct": 80,
+        "paused": False,
+    },
+}
+
+# Global event queue — re-created in startup_event() to match uvicorn's loop
+event_queue: asyncio.Queue = asyncio.Queue()
